Split up and clean up the IPCRingBuffer implementation.
